+++ crates/nostr/src/nips/nip57.rs
@@ -351,7 +351,7 @@ where
     R: RngCore,
     T: AsRef<[u8]>,
 {
-    let key: [u8; 32] = util::generate_shared_key(secret_key, public_key);
+    let key: [u8; 32] = util::generate_shared_key(secret_key, public_key)?;
     let mut iv: [u8; 16] = [0u8; 16];
     rng.fill_bytes(&mut iv);
 
@@ -386,7 +386,7 @@ pub fn decrypt_sent_private_zap_message(
     // Re-create our ephemeral encryption key
     let secret_key: SecretKey =
         create_encryption_key(secret_key, public_key, private_zap_event.created_at)?;
-    let key: [u8; 32] = util::generate_shared_key(&secret_key, public_key);
+    let key: [u8; 32] = util::generate_shared_key(&secret_key, public_key)?;
 
     // decrypt like normal
     decrypt_private_zap_message(key, private_zap_event)
@@ -398,7 +398,7 @@ pub fn decrypt_received_private_zap_message(
     secret_key: &SecretKey,
     private_zap_event: &Event,
 ) -> Result<Event, Error> {
-    let key: [u8; 32] = util::generate_shared_key(secret_key, &private_zap_event.pubkey);
+    let key: [u8; 32] = util::generate_shared_key(secret_key, &private_zap_event.pubkey)?;
     decrypt_private_zap_message(key, private_zap_event)
 }